09/00 create addCSVToMap()

+++ cpp_module_09/ex00/BitcoinExchange.hpp
@@ -5,6 +5,7 @@
 #include <string>
 #include <map>
 #include <fstream>
+#include <sstream>
 
 class BitcoinExchange {
 	public:
@@ -16,8 +17,8 @@ class BitcoinExchange {
 		static bool	canOpenFiles(const char *arg_file);
 		void	addCSVToBitcoinMap();
 	private:
-		std::map<std::string, size_t> _data_map;
-		// std::map<std::string, size_t> _input_map;
+		std::map<std::string, int> _data_map;
+		// std::map<std::string, int> _input_map;
 };
 
 #endif
